checking if i rename to deleted file if go delete

+++ MRManager/MRManager-UnitTests/Process-01-SystemIntialization.cs
@@ -4,24 +4,18 @@ using System.Collections.Generic;
 using System.IO;
 using System.Reactive.Linq;
 using System.Threading;
-using System.Threading.Tasks;
 using SystemInterfaces;
 using Actor.Interfaces;
 using Common;
-using CommonMessages;
-using DataEntites;
 using DataServices.Actors;
 using Domain.Interfaces;
 using EF.DBContexts;
 using EF.Entities;
 using EventAggregator;
-using EventMessages;
 using Interfaces;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using RevolutionEntities.Process;
-using StartUp.Messages;
 using Utilities;
-using ViewMessages;
 using ViewModel.Interfaces;
 using ViewModels;
 using RevolutionLogger;
@@ -36,47 +30,36 @@ namespace MRManager_UnitTests
     {
 
         public static ISystemSource Source => new Source(Guid.NewGuid(), "TestCase" + typeof(SystemIntializationProcess).GetFriendlyName(),new SourceType(typeof(SystemIntializationProcess)), new MachineInfo(Environment.MachineName, Environment.ProcessorCount));
-        public void StartSystem()
+        private static bool started;
+
+        public static void StartSystem()
         {
+            if (started) return;
+            started = true;
             if (File.Exists("MRManager-TEST-Logs.xml")) File.Delete("MRManager-TEST-Logs.xml");
             Logger.Initialize();
             var t = new MRManagerDBContext().GetType().Assembly;
             var x = new EFEntity<IEntity>().GetType().Assembly;
             BootStrapper.BootStrapper.Instance.StartUp(t, x, false);
             var mainWindow = MainWindowViewModel.Instance;
-
+           
         }
 
 
+        // Decide not to Continuous Test Class cuz instaniation and testing to coupled
         [TestMethod]
-        public void SystemIntialization()
+        public void SystemTest()
         {
             RegisterProcess1Events();
+           
             StartSystem();
-            Thread.Sleep(TimeSpan.FromSeconds(30));
+            
+            
+            Thread.Sleep(TimeSpan.FromSeconds(15));
+
             Process1Asserts();
-            AssertProcessLogs();
+           
         }
-
-
-        public static void AssertProcessLogs()
-        {
-            foreach (var p in startedProcesses.ToArray())
-            {
-                EventMessageBus.Current.Publish(
-                    new RequestProcessLog(
-                        new StateCommandInfo(p.Process.Id, RevolutionData.Context.Process.Commands.CreateLog), p.Process, Source),
-                    Source);
-            }
-
-            Thread.Sleep(TimeSpan.FromSeconds(5));
-            Assert.IsTrue(ProcessLogs.Count == startedProcesses.Count);
-        }
-
-        private IProcessSystemMessage process2ServiceActorStarted;
-
-
-       
 
 
 
@@ -88,15 +71,21 @@ namespace MRManager_UnitTests
         private static IProcessSystemMessage screenViewModelLoadedInMainWindowViewModel;
         private static IProcessSystemMessage processCompleted;
        // private IProcessSystemMessage mainWindowViewModelCreated;
-
-        public static ConcurrentQueue<IProcessEventFailure> EventFailures = new ConcurrentQueue<IProcessEventFailure>();
-        private static ConcurrentQueue<ISystemProcessStarted> startedProcesses = new ConcurrentQueue<ISystemProcessStarted>();
-    
+        private IProcessSystemMessage process2Started;
+        private IViewModelCreated<ISigninViewModel> LoginViewModelCreated;
+        private IViewModelLoaded<IScreenModel, IViewModel> LoginViewModelLoadedInMScreenViewModel;
+        private IRequestProcessState process2StateRequest;
+        private List<IProcessStateMessage<ISignInInfo>> process2StateMessageList = new List<IProcessStateMessage<ISignInInfo>>();
+        private IGetEntityViewWithChanges<ISignInInfo> UserNameEntityChanges;
+        private IEntityViewWithChangesFound<ISignInInfo> userFound;
+        private IUserValidated userValidated;
+        private IServiceStarted<IEntityViewDataServiceActor<IGetEntityViewWithChanges<ISignInInfo>>> getEntityChangesActor;
+        private static ConcurrentQueue<IProcessEventFailure> EventFailures = new ConcurrentQueue<IProcessEventFailure>();
+        private IViewStateLoaded<ISigninViewModel, IProcessState<ISignInInfo>> viewLoadedState;
 
 
         public static void RegisterProcess1Events()
         {
-            EventMessageBus.Current.GetEvent<ISystemProcessStarted>(Source).Subscribe(x => startedProcesses.Enqueue(x));
             EventMessageBus.Current.GetEvent<IProcessLogCreated>(Source).Subscribe(x => ProcessLogs.Enqueue(x));
             EventMessageBus.Current.GetEvent<IComplexEventLogCreated>(Source).Subscribe(x => ComplextEventLogs.Enqueue(x));
             EventMessageBus.Current.GetEvent<IProcessEventFailure>(Source).Subscribe(x => EventFailures.Enqueue(x));
@@ -112,9 +101,10 @@ namespace MRManager_UnitTests
             EventMessageBus.Current.GetEvent<ISystemProcessCompleted>(Source).Where(procesPredicate).Subscribe(x => processCompleted = x);
         }
 
-        public static ConcurrentQueue<IComplexEventLogCreated> ComplextEventLogs = new ConcurrentQueue<IComplexEventLogCreated>();
+        private static ConcurrentQueue<IComplexEventLogCreated> ComplextEventLogs = new ConcurrentQueue<IComplexEventLogCreated>();
 
-        public static ConcurrentQueue<IProcessLogCreated> ProcessLogs =new ConcurrentQueue<IProcessLogCreated>();
+        private static ConcurrentQueue<IProcessLogCreated> ProcessLogs =new ConcurrentQueue<IProcessLogCreated>();
+       
 
         public static void Process1Asserts()
         {
